+++ plugins/auth-backend/src/providers/atlassian/provider.ts
@@ -58,6 +58,10 @@ export const atlassianDefaultAuthHandler: AuthHandler<OAuthResult> = async ({
   profile: makeProfileInfo(fullProfile, params.id_token),
 });
 
+/**
+ * @public
+ * @deprecated This export is deprecated and will be removed in the future.
+ */
 export class AtlassianAuthProvider implements OAuthHandlers {
   private readonly _strategy: AtlassianStrategy;
   private readonly signInResolver?: SignInResolver<OAuthResult>;
@@ -162,25 +166,6 @@ export class AtlassianAuthProvider implements OAuthHandlers {
 }
 
 /**
- * @public
- * @deprecated This type has been inlined into the create method and will be removed.
- */
-export type AtlassianProviderOptions = {
-  /**
-   * The profile transformation function used to verify and convert the auth response
-   * into the profile that will be presented to the user.
-   */
-  authHandler?: AuthHandler<OAuthResult>;
-
-  /**
-   * Configure sign-in for this provider, without it the provider can not be used to sign users in.
-   */
-  signIn?: {
-    resolver: SignInResolver<OAuthResult>;
-  };
-};
-
-/**
  * Auth provider integration for atlassian auth
  *
  * @public
@@ -230,9 +215,3 @@ export const atlassian = createAuthProviderIntegration({
       });
   },
 });
-
-/**
- * @public
- * @deprecated Use `providers.atlassian.create` instead
- */
-export const createAtlassianProvider = atlassian.create;