+++ plugins/auth-backend/src/providers/saml/index.ts
@@ -14,8 +14,5 @@
  * limitations under the License.
  */
 
-export {
-  createSamlProvider,
-  samlNameIdEntityNameSignInResolver,
-} from './provider';
-export type { SamlProviderOptions, SamlAuthResult } from './provider';
+export { saml } from './provider';
+export type { SamlAuthResult } from './provider';