+++ packages/dev-utils/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @backstage/dev-utils
 
+## 1.0.4-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.10.0-next.2
+  - @backstage/catalog-model@1.1.0-next.2
+  - @backstage/theme@0.2.16-next.1
+  - @backstage/plugin-catalog-react@1.1.2-next.2
+  - @backstage/app-defaults@1.0.4-next.2
+  - @backstage/integration-react@1.1.2-next.2
+
+## 1.0.4-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.9.6-next.1
+  - @backstage/catalog-model@1.1.0-next.1
+  - @backstage/theme@0.2.16-next.0
+  - @backstage/app-defaults@1.0.4-next.1
+  - @backstage/integration-react@1.1.2-next.1
+  - @backstage/test-utils@1.1.2-next.1
+  - @backstage/plugin-catalog-react@1.1.2-next.1
+
 ## 1.0.4-next.0
 
 ### Patch Changes