+++ plugins/api-docs/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @backstage/plugin-api-docs
 
+## 0.8.7-next.2
+
+### Patch Changes
+
+- a274fe38b9: Add hidden title column to catalog and API table to enable filtering by title.
+- Updated dependencies
+  - @backstage/core-components@0.10.0-next.2
+  - @backstage/catalog-model@1.1.0-next.2
+  - @backstage/theme@0.2.16-next.1
+  - @backstage/plugin-catalog@1.4.0-next.2
+  - @backstage/plugin-catalog-react@1.1.2-next.2
+
+## 0.8.7-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.9.6-next.1
+  - @backstage/catalog-model@1.1.0-next.1
+  - @backstage/plugin-catalog@1.3.1-next.1
+  - @backstage/theme@0.2.16-next.0
+  - @backstage/plugin-catalog-react@1.1.2-next.1
+
 ## 0.8.7-next.0
 
 ### Patch Changes