+++ plugins/bazaar/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @backstage/plugin-bazaar
 
+## 0.1.22-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.10.0-next.2
+  - @backstage/catalog-model@1.1.0-next.2
+  - @backstage/plugin-catalog@1.4.0-next.2
+  - @backstage/cli@0.18.0-next.2
+  - @backstage/plugin-catalog-react@1.1.2-next.2
+
+## 0.1.22-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/core-components@0.9.6-next.1
+  - @backstage/cli@0.18.0-next.1
+  - @backstage/catalog-model@1.1.0-next.1
+  - @backstage/plugin-catalog@1.3.1-next.1
+  - @backstage/catalog-client@1.0.4-next.1
+  - @backstage/plugin-catalog-react@1.1.2-next.1
+
 ## 0.1.22-next.0
 
 ### Patch Changes