+++ plugins/permission-node/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-permission-node
 
+## 0.6.3-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.14.1-next.1
+  - @backstage/errors@1.1.0-next.0
+  - @backstage/plugin-auth-node@0.2.3-next.1
+  - @backstage/plugin-permission-common@0.6.3-next.0
+
 ## 0.6.3-next.0
 
 ### Patch Changes